Add a lookup for attribute given an rdf predicate
In an ingest script I was reading metadata from a graph of triples and needed to derive the attribute name from the predicate describing the data.  After some misunderstanding (`rdf_annotation_for_attr` will find the predicate given the attribute) we discovered that this wasn't something we'd thought about.  So here it is!

Usage: `instance.attribute_for_rdf_annotation(predicate)`